+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### New
 
+- Added `parent` Organization Field to the Organization Model.
 - Added sending Email to the Team from the "Contact Us" Page.
 - Added the Features Feature for toggling the new testable Features.